fix: move the loader to the components folder
